pyproject.toml
[project]
name = "cemclrn-mp-train"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "fastapi>=0.122.0",
    "matplotlib>=3.10.7",
    "mysql-connector-python>=9.5.0",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "pytorch-lightning>=2.5.6",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
